Refactor and clean-up module beans init
